backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
toml==0.10.2
openai==1.6.1
